fix: move error logic to handleCloudFormationError (rename file)
